add trait Timestampable in Dream & Resource and Faq
