+++ src/AppBundle/Document/Timestampable.php
@@ -9,10 +9,6 @@ use JMS\Serializer\Annotation\Type;
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
 
 /**
- * Traits Timestampable
- * @package AppBundle\Document
- *
- * @ODM\Document(collection="trait")
  * @ExclusionPolicy("all")
  */
 
@@ -46,4 +42,73 @@ trait Timestampable
      * @Type("DateTime")
      */
     protected $deletedAt;
+
+    /**
+     * Set createdAt
+     *
+     * @param  date $createdAt
+     * @return self
+     */
+    public function setCreatedAt($createdAt)
+    {
+        $this->createdAt = $createdAt;
+
+        return $this;
+    }
+
+    /**
+     * Get createdAt
+     *
+     * @return date $createdAt
+     */
+    public function getCreatedAt()
+    {
+        return $this->createdAt;
+    }
+
+    /**
+     * Set updatedAt
+     *
+     * @param  date $updatedAt
+     * @return self
+     */
+    public function setUpdatedAt($updatedAt)
+    {
+        $this->updatedAt = $updatedAt;
+
+        return $this;
+    }
+
+    /**
+     * Get updatedAt
+     *
+     * @return date $updatedAt
+     */
+    public function getUpdatedAt()
+    {
+        return $this->updatedAt;
+    }
+
+    /**
+     * Set deletedAt
+     *
+     * @param  date $deletedAt
+     * @return self
+     */
+    public function setDeletedAt($deletedAt)
+    {
+        $this->deletedAt = $deletedAt;
+
+        return $this;
+    }
+
+    /**
+     * Get deletedAt
+     *
+     * @return date $deletedAt
+     */
+    public function getDeletedAt()
+    {
+        return $this->deletedAt;
+    }
 }
